quickSortKernel(arr, low, pivot_index-1);// pivot-1 can be negative(0-1), using long type not unsiged int

+++ cpp/src/sort/quick_sort.h
@@ -7,13 +7,13 @@ namespace SORT{
     namespace QUICK_SORT{
 
     template<typename T>
-    unsigned int partition(T arr[], unsigned int low, unsigned int high);
+    long partition(T arr[], long low, long high);
 
     template<typename T>
-    void quickSortKernel(T arr[], unsigned int n);
+    void quickSortKernel(T arr[], long n);
 
     template<typename T>
-    void quickSortKernel(T arr[], unsigned int low, unsigned int high);
+    void quickSortKernel(T arr[], long low, long high);
 
     }
 }
