+++ python/AlgoLibR/sort/csort.pxd
@@ -8,7 +8,7 @@ cdef extern from "sort/bubble_sort.h" namespace "SORT::BUBBLE_SORT":
     cdef void bubbleSortKernel[T](T *arr, int n)
 
 cdef extern from "sort/quick_sort.h" namespace "SORT::QUICK_SORT":
-    cdef void quickSortKernel[T](T *arr, unsigned int n)
+    cdef void quickSortKernel[T](T *arr, long n)
 
 cdef extern from "sort/selection_sort.h" namespace "SORT::SELECTION_SORT":
     cdef void selectionSortKernel[T](T *arr, unsigned int n)
